employees data provider 2

+++ src/main/java/org/ok/bella/ui/employees/ClientEmployeesDataProvider.java
@@ -1,7 +1,6 @@
 package org.ok.bella.ui.employees;
 
 import lombok.extern.slf4j.Slf4j;
-import org.springframework.context.annotation.Primary;
 import org.springframework.stereotype.Service;
 
 @Service
@@ -9,7 +8,7 @@ import org.springframework.stereotype.Service;
 public class ClientEmployeesDataProvider implements EmployeesDataProvider {
 
     @Override
-    public int count() {
+    public long count() {
         return 31;
     }
 }
